refactor - change bottom bar package - PrimaryButton shadow shape
